+++ projects/ngx-xchange-icons/CHANGELOG.md
@@ -1,3 +1,9 @@
+# 0.1.4 24/02/2025
+* (add) - Navigator table icon
+* (add) - Navigator board icon
+* (add) - navigator graph
+* (add) - new icons to public file
+
 # 0.1.3 14/02/2025
 * (change) - icon clock size and color