+++ projects/ngx-xchange-ui/CHANGELOG.md
@@ -1,9 +1,12 @@
-# 0.1.7 19/02/2025
+# 0.1.7 24/02/2025
 * (add) - Navigator objective card component
 * (add) - Navigator toggle button component
 * (add) - Navigator status filter component
 * (change) - styles colors for toggle buttons
 * (change) - change library version and change log file
+* (add) - Navigator filter selector switch component
+* (add) - Navigator filter type selector component
+* (add) - Components to public file
 
 # 0.1.6 17/08/2025
 * (change) - boder radius on objective card